Fix TestSystemLogsCanUseLogInput and run it on all OSes
Fix TestSystemLogsCanUseLogInput and move it to a file without the
linux build constraint so it can run on all OSes supported by the
system integration.
